Rename tabs directory to pages and clarify meal plan components
