+++ src/components/pages/Training.tsx
